move IDTreeS notebooks to dedicated dir, move cutout script to script folder and generalize it a bit
